+++ packages/brick_build/lib/src/serdes_generator.dart
@@ -266,7 +266,14 @@ abstract class SerdesGenerator<_FieldAnnotation extends FieldSerializable,
       FieldElement field, _FieldAnnotation annotation, SharedChecker<Model> checker) {
     final isComputedGetter = FieldsForClass.isComputedGetter(field);
 
-    return (isComputedGetter && doesDeserialize) || annotation.ignore || !checker.isSerializable;
+    if (isComputedGetter && doesDeserialize) return true;
+    if (annotation.ignore) return true;
+
+    final hasGenerator =
+        doesDeserialize ? annotation.fromGenerator != null : annotation.toGenerator != null;
+    if (!checker.isSerializable && hasGenerator) return false;
+
+    return !checker.isSerializable;
   }
 
   /// The field's name when being serialized to a provider. Optionally, a checker can reveal