+++ packages/brick_rest_generators/test/rest_model_serdes_generator_test.dart
@@ -8,6 +8,8 @@ import 'package:brick_build/testing.dart';
 
 import 'rest_model_serdes_generator/test_enum_as_string.dart' as enum_as_string;
 import 'rest_model_serdes_generator/test_ignore_from_to.dart' as ignore_from_to;
+import 'rest_model_serdes_generator/test_unserializable_field_with_generator.dart'
+    as unserializable_field_with_generator;
 
 final _generator = TestGenerator();
 final folder = 'rest_model_serdes_generator';
@@ -22,6 +24,11 @@ void main() {
 
       test('ignoreFrom ignoreTo', () async {
         await generateExpectation('ignore_from_to', ignore_from_to.output);
+      });
+
+      test('fromGenerator toGenerator', () async {
+        await generateExpectation(
+            'unserializable_field_with_generator', unserializable_field_with_generator.output);
       });
     });
   });