+++ packages/brick_sqlite_generators/lib/src/sqlite_serdes_generator.dart
@@ -5,6 +5,7 @@ import 'package:brick_sqlite_abstract/annotations.dart';
 import 'package:brick_sqlite_abstract/db.dart' show InsertForeignKey;
 import 'package:brick_sqlite_abstract/sqlite_model.dart';
 import 'package:meta/meta.dart';
+import 'package:source_gen/source_gen.dart';
 
 import 'sqlite_fields.dart';
 
@@ -32,5 +33,27 @@ abstract class SqliteSerdesGenerator<_Model extends SqliteModel>
     }
 
     return annotatedName;
+  }
+
+  @override
+  bool ignoreCoderForField(field, annotation, checker) {
+    if (annotation.columnType != null) {
+      if (checker.isSerializable) return false;
+
+      if (doesDeserialize) {
+        if (annotation.fromGenerator == null) {
+          throw InvalidGenerationSourceError(
+              'Decalaring column type ${annotation.columnType} on ${field.name} requires `fromGenerator` to be declared');
+        }
+      } else {
+        if (annotation.toGenerator == null) {
+          throw InvalidGenerationSourceError(
+              'Decalaring column type ${annotation.columnType} on ${field.name} requires `toGenerator` to be declared');
+        }
+      }
+
+      return false;
+    }
+    return super.ignoreCoderForField(field, annotation, checker);
   }
 }