+++ packages/brick_sqlite_generators/test/migration_generator/test_from_identical_schema.dart
@@ -16,7 +16,8 @@ class Migration1 extends Migration {
 
 final schema = Schema(2, generatorVersion: 1, tables: <SchemaTable>{
   SchemaTable('User', columns: <SchemaColumn>{
-    SchemaColumn('_brick_id', int, autoincrement: true, nullable: false, isPrimaryKey: true),
-    SchemaColumn('name', String),
+    SchemaColumn('_brick_id', Column.integer,
+        autoincrement: true, nullable: false, isPrimaryKey: true),
+    SchemaColumn('name', Column.varchar),
   })
 });