+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_column_type_without_generator.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+import 'package:brick_sqlite_abstract/annotations.dart';
+import 'package:brick_sqlite_abstract/sqlite_model.dart';
+
+@SqliteSerializable()
+class ColumnTypeWithoutGenerator extends SqliteModel {
+  @Sqlite(columnType: Column.Double, toGenerator: '%INSTANCE_PROPERTY%')
+  final File field;
+
+  ColumnTypeWithoutGenerator(this.field);
+}