+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator_test.dart
@@ -6,6 +6,7 @@ import 'package:test/test.dart';
 import 'package:source_gen/source_gen.dart';
 import 'package:brick_build/testing.dart';
 
+import 'sqlite_model_serdes_generator/test_sqlite_column_type.dart' as _$sqliteColumnType;
 import 'sqlite_model_serdes_generator/test_sqlite_unique.dart' as _$sqliteUnique;
 import 'sqlite_model_serdes_generator/test_field_with_type_argument.dart'
     as _$fieldWithTypeArgument;
@@ -37,9 +38,21 @@ void main() {
           throwsA(TypeMatcher<InvalidGenerationSourceError>()),
         );
       });
+
+      test('ColumnTypeWithoutFrom', () async {
+        final reader = await generateReader('column_type_without_generator');
+        expect(
+          () async => await _generator.generate(reader, null),
+          throwsA(TypeMatcher<InvalidGenerationSourceError>()),
+        );
+      });
     });
 
     group('@Sqlite', () {
+      test('columnType', () async {
+        await generateAdapterExpectation('sqlite_column_type', _$sqliteColumnType.output);
+      });
+
       test('unique', () async {
         await generateAdapterExpectation('sqlite_unique', _$sqliteUnique.output);
       });