+++ packages/brick_sqlite_generators/test/sqlite_schema/test_index_annotation.dart
@@ -14,10 +14,10 @@ final Set<Migration> migrations = <Migration>{};
 /// A consumable database structure including the latest generated migration.
 final schema = Schema(0, generatorVersion: 1, tables: <SchemaTable>{
   SchemaTable('IndexAnnotation', columns: <SchemaColumn>{
-    SchemaColumn('_brick_id', int,
+    SchemaColumn('_brick_id', Column.integer,
         autoincrement: true, nullable: false, isPrimaryKey: true),
-    SchemaColumn('non_unique', String),
-    SchemaColumn('unique', String, unique: true)
+    SchemaColumn('non_unique', Column.varchar),
+    SchemaColumn('unique', Column.varchar, unique: true)
   }, indices: <SchemaIndex>{
     SchemaIndex(columns: ['non_unique'], unique: false),
     SchemaIndex(columns: ['unique'], unique: true)