+++ packages/brick_sqlite_generators/test/sqlite_schema/test_one_to_one_association.dart
@@ -24,18 +24,18 @@ final Set<Migration> migrations = <Migration>{};
 /// A consumable database structure including the latest generated migration.
 final schema = Schema(0, generatorVersion: 1, tables: <SchemaTable>{
   SchemaTable('SqliteAssoc', columns: <SchemaColumn>{
-    SchemaColumn('_brick_id', int,
+    SchemaColumn('_brick_id', Column.integer,
         autoincrement: true, nullable: false, isPrimaryKey: true)
   }, indices: <SchemaIndex>{}),
   SchemaTable('OneToOneAssocation', columns: <SchemaColumn>{
-    SchemaColumn('_brick_id', int,
+    SchemaColumn('_brick_id', Column.integer,
         autoincrement: true, nullable: false, isPrimaryKey: true),
-    SchemaColumn('assoc_SqliteAssoc_brick_id', int,
+    SchemaColumn('assoc_SqliteAssoc_brick_id', Column.integer,
         isForeignKey: true,
         foreignTableName: 'SqliteAssoc',
         onDeleteCascade: false,
         onDeleteSetDefault: false),
-    SchemaColumn('assoc2_SqliteAssoc_brick_id', int,
+    SchemaColumn('assoc2_SqliteAssoc_brick_id', Column.integer,
         isForeignKey: true,
         foreignTableName: 'SqliteAssoc',
         onDeleteCascade: false,