+++ packages/brick_sqlite_generators/test/sqlite_schema/test_simple.dart
@@ -55,9 +55,9 @@ final Set<Migration> migrations = <Migration>{};
 /// A consumable database structure including the latest generated migration.
 final schema = Schema(0, generatorVersion: 1, tables: <SchemaTable>{
   SchemaTable('Simple', columns: <SchemaColumn>{
-    SchemaColumn('_brick_id', int,
+    SchemaColumn('_brick_id', Column.integer,
         autoincrement: true, nullable: false, isPrimaryKey: true),
-    SchemaColumn('name', String)
+    SchemaColumn('name', Column.varchar)
   }, indices: <SchemaIndex>{})
 });
 ''';